Remove the option for customized pycodestyle configuration (#5)
When using codecheck in the CI for https://github.com/yugabyte/compiler-identification, somehow the custom pycodestyle.ini configuration file did not work, even though it was properly specified on the pycodestyle command line using `--config`. The tox.ini technique seems to always work, so making that the default approach for customizing the code style.